+++ src/main/java/com/airbnl/managerservice/service/Interfaces/IHotelService.java
@@ -6,8 +6,8 @@ import java.util.List;
 
 public interface IHotelService {
     Hotel save(Hotel hotel);
-    List<Hotel> getAllByManagerID(int managerId);
-    Hotel getByHotelIdAndManagerId(int hotelId, int managerId);
+    List<Hotel> getAllByManagerID(Long managerId);
+    Hotel getByHotelIdAndManagerId(Long hotelId, Long managerId);
     Hotel update(Hotel hotel);
-    Hotel deleteByIdAndManagerID(int hotelId,int managerId);
+    Hotel deleteByIdAndManagerID(Long hotelId,Long managerId);
 }
